DRY up employment forms
