+++ app/views/jobseekers/job_applications/employments/new.html.slim
@@ -1,11 +1,11 @@
-- content_for :page_title_prefix, job_application_page_title_prefix(form, t(".title"))
+- content_for :page_title_prefix, job_application_page_title_prefix(@form, t(".title"))
 
 .govuk-grid-row
   .govuk-grid-column-two-thirds
     span.govuk-caption-l = t(".caption")
     h1.govuk-heading-xl = t(".heading")
 
-    = form_for form, url: jobseekers_job_application_employments_path(job_application), method: :post do |f|
+    = form_for @form, url: jobseekers_job_application_employments_path(job_application), method: :post do |f|
       = f.govuk_error_summary
 
       = render "employment_fields", f: f
